refactor: ordertable 패키지 변경
