+++ src/main/java/kitchenpos/dao/OrderTableDao.java
@@ -1,9 +1,8 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.OrderTable;
-
 import java.util.List;
 import java.util.Optional;
+import kitchenpos.domain.ordertable.OrderTable;
 
 public interface OrderTableDao {
     OrderTable save(OrderTable entity);
