+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -11,9 +11,9 @@ import kitchenpos.dao.TableGroupDao;
 import kitchenpos.dao.fakedao.InMemoryOrderDao;
 import kitchenpos.dao.fakedao.InMemoryOrderTableDao;
 import kitchenpos.dao.fakedao.InMemoryTableGroupDao;
-import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.OrderTableFactory;
 import kitchenpos.domain.TableGroupFactory;
+import kitchenpos.domain.ordertable.OrderTable;
 import org.assertj.core.api.ThrowableAssert.ThrowingCallable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
