+++ src/test/java/kitchenpos/dao/fakedao/InMemoryOrderTableDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import kitchenpos.dao.OrderTableDao;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.ordertable.OrderTable;
 
 public class InMemoryOrderTableDao implements OrderTableDao {
